feat: support setting environment variables
