+++ src/types/common.ts
@@ -1,0 +1,11 @@
+export interface ResponseError {
+  code?: string;
+  message?: string;
+  items?: string[];
+}
+
+export interface ResponseDto<T = unknown> {
+  ok: boolean;
+  result?: T;
+  error?: ResponseError;
+}
