+++ src/types/deployment/deployment.ts
@@ -1,0 +1,15 @@
+export interface Disk {
+  name: string;
+  mountPath: string;
+  subPath: string;
+}
+
+interface ResourceValue {
+  cpu: string;
+  memory: string;
+}
+
+export interface Resource {
+  requests: ResourceValue;
+  limits: ResourceValue;
+}
