+++ src/types/errors.ts
@@ -1,0 +1,4 @@
+export enum ErrorCode {
+  DEPLOYMENT_NOT_FOUND = 'DEPLOYMENT_NOT_FOUND',
+  DEPLOYMENT_NAME_ALREADY_EXISTS = 'DEPLOYMENT_NAME_ALREADY_EXISTS',
+}
